32/64 bit changes to Epetra, FEApp, stokhos, phalanx.
Include Epetra_ConfigDefs.h before any other file in those
Epetra files that make use of EPETRA_NO_64BIT_GLOBAL_INDICES
or EPETRA_NO_32BIT_GLOBAL_INDICES macros.  Avoids vtable
inconsistency (at least) across translation units.

Remove the recently introduced unsigned int/long long based
constructors and functions from Map classes.  They were
introduced to allow overloading in external packages so that
ambiguity between int and long long does not happen but they
lead to a different overloading ambiguity when code is compiled
in single mode (32 or 64).  The correct fix is below.

Cast argument to Epetra_LocalMap to be always int.  It will
then create int map when 32 bit indices are present and long
long map when they are not.  This avoids ambiguity between
int and long long when size_t or unsigned int is passed.

Cast argument to LID() to be int or long long depending on
compilation mode.

Keep some long long code that was always available inside
but trying to reduce clutter in object code and so that
linker/compiler errors can happen when they should.

Explicitly exclude the Epetra_LongLong*/ h/cpp files when
long long indices are not present even though the
translation unit would be empty.  Reduce cmake clutter.

When a class contains foo_int_ and foo_LL_ data members,
make them available depending on compilation mode.  Saves
sizeof(object) and hopefully leads to linker/compiler errors
when they should happen.

+++ packages/epetra/test/CrsGraph_LL/cxx_main.cpp
@@ -549,7 +549,7 @@ int checkCopyAndAssignment(Epetra_Comm& Comm, bool verbose) {
 
 	// create initial Map and Graph
 	const int NumIndicesPerRow = 10;
-	const int NumGlobalElements = 50;
+	const long long NumGlobalElements = 50;
 	const int IndexBase = 0;
 	Epetra_Map Map1(NumGlobalElements, IndexBase, Comm);
 	Epetra_CrsGraph Graph1(Copy, Map1, NumIndicesPerRow);
